pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "ogmac"
version = "0.1.0"
description = "Outlook to Google calendar one-way sync for macOS"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "ogmac contributors" }]
requires-python = ">=3.11"
keywords = ["calendar", "outlook", "google-calendar", "macos", "launchd", "sync", "eventkit"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Environment :: MacOS X",
  "Intended Audience :: End Users/Desktop",
  "License :: OSI Approved :: MIT License",
  "Operating System :: MacOS :: MacOS X",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Office/Business :: Scheduling",
  "Topic :: Utilities",
]
dependencies = [
  "msal>=1.28",
  "google-auth>=2.29",
  "google-auth-oauthlib>=1.2",
  "google-api-python-client>=2.130",
  "httpx>=0.27",
  "keyring>=25.0",
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "pyobjc-framework-EventKit>=10.0; sys_platform == 'darwin'",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-httpx>=0.30",
  "pytest-cov>=5.0",
  "ruff>=0.4",
  "mypy>=1.10",
]

[project.urls]
Homepage = "https://github.com/razbahri/ogmac"
Repository = "https://github.com/razbahri/ogmac"
Issues = "https://github.com/razbahri/ogmac/issues"
Changelog = "https://github.com/razbahri/ogmac/blob/main/CHANGELOG.md"

[project.scripts]
ogmac = "ogmac.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
